Rename to deputy classes with named deputy in the title
