+++ api/app/src/Entity/Deputy.php
@@ -9,15 +9,15 @@ use Doctrine\ORM\Mapping as ORM;
 use JMS\Serializer\Annotation as JMS;
 
 /**
- * Named Deputy.
+ * Deputy.
  *
  * @ORM\Table(name="named_deputy", indexes={@ORM\Index(name="named_deputy_uid_idx", columns={"deputy_uid"})})
  *
- * @ORM\Entity(repositoryClass="App\Repository\NamedDeputyRepository")
+ * @ORM\Entity(repositoryClass="DeputyRepository")
  *
  * @ORM\HasLifecycleCallbacks()
  */
-class NamedDeputy
+class Deputy
 {
     use CreateUpdateTimestamps;
