+++ client/app/tests/phpunit/EventSubscriber/DeputyChangedOrgSubscriberTest.php
@@ -9,7 +9,7 @@ use App\EventSubscriber\DeputyChangedOrgSubscriber;
 use App\Service\Audit\AuditEvents;
 use App\Service\Time\DateTimeProvider;
 use App\TestHelpers\ClientHelpers;
-use App\TestHelpers\NamedDeputyHelper;
+use App\TestHelpers\DeputyHelper;
 use App\TestHelpers\OrganisationHelpers;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
@@ -43,7 +43,7 @@ class DeputyChangedOrgSubscriberTest extends TestCase
         //      Client record currently in database
         $client = ClientHelpers::createClient();
         $clientOrg = OrganisationHelpers::createActivatedOrganisation();
-        $namedDeputy = NamedDeputyHelper::createNamedDeputy();
+        $namedDeputy = DeputyHelper::createDeputy();
         $client->setDeputy($namedDeputy);
         $client->setOrganisation($clientOrg);
